LoginとReceiptとStockチェックを追加する

+++ src/main/java/com/example/onlineshoppingsystem/repository/ProductSizeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductSizeRepository extends JpaRepository <ProductSize , Long>{
+public interface ProductSizeRepository extends JpaRepository <ProductSize , Integer>{
     
 }
